+++ code/backend/src/controllers/auth.controller.js
@@ -83,7 +83,7 @@ const login = asyncHandler(async (req, res) => {
 
     res.cookie('token', token, {
         httpOnly: true,
-        secure: false,        
+        secure: false,        // ต้อง true เพราะ https
         sameSite: 'none',    // อนุญาต cross-site
         maxAge: 7 * 24 * 60 * 60 * 1000
     });