+++ code/backend/src/routes/index.js
@@ -7,8 +7,11 @@ const driverVerifRoutes = require('./driverVerification.routes');
 const bookingRoutes = require('./booking.routes');
 const notificationRoutes = require('./notification.routes')
 const mapRoutes = require('./maps.routes')
-const { prisma } = require("../utils/prisma"); // adjust path if needed
+const monitorRoutes = require('./monitor.routes');
+const blacklistRoutes = require('./blacklist.routes');
+const reportRoutes = require('./report.routes');
 
+const { prisma } = require("../utils/prisma"); // adjust path if needed
 
 const router = express.Router();
 
@@ -19,14 +22,15 @@ router.use('/routes', routeRoutes);
 router.use('/driver-verifications', driverVerifRoutes);
 router.use('/bookings', bookingRoutes);
 router.use('/notifications', notificationRoutes);
-router.use('/api/maps', mapRoutes);
+router.use('/api/maps', mapRoutes); 
 
-//เพิ่ม monitor routes
-const monitorRoutes = require('./monitor.routes');
+// monitor routes
 router.use('/monitor', monitorRoutes);
 
-// เพิ่ม blacklist routes
-const blacklistRoutes = require('./blacklist.routes');
+// blacklist routes
 router.use('/blacklists', blacklistRoutes);
+
+// report routes
+router.use('/reports', reportRoutes);
 
 module.exports = router;